stackoverflow/requirements.txt
pandas-gbq==0.18.1
google-cloud-logging